feat: ApiSchemaRouter, DidDocRouter
