+++ packages/daf-cli/src/server.ts
@@ -2,12 +2,13 @@ import express from 'express'
 import program from 'commander'
 import ngrok from 'ngrok'
 import parse from 'url-parse'
-import { AgentRouter, AgentApiSchemaRouter, AgentDidDocRouter, didDocEndpoint } from 'daf-express'
+import { AgentRouter, ApiSchemaRouter, DidDocRouter, didDocEndpoint } from 'daf-express'
 import swaggerUi from 'swagger-ui-express'
 import { getAgent, getConfig } from './setup'
 import { createObjects } from './lib/objectCreator'
 import passport from 'passport'
 import Bearer from 'passport-http-bearer'
+import { IIdentity } from 'daf-core'
 const exphbs = require('express-handlebars')
 const hbs = exphbs.create({
   helpers: {
@@ -26,13 +27,30 @@ program
     const agent = getAgent(program.config)
     const { server: options } = createObjects(getConfig(program.config), { server: '/server' })
 
-    const exposedMethods = options.exposedMethods ? options.exposedMethods : agent.availableMethods()
+    /**
+     * Ngrok configuration
+     */
+    let baseUrl = options.baseUrl
+    if (options.ngrok?.connect) {
+      baseUrl = await ngrok.connect({
+        addr: cmd.port || options.port,
+        subdomain: options.ngrok.subdomain,
+        region: options.ngrok.region,
+        authtoken: options.ngrok.authtoken,
+      })
+      app.set('trust proxy', 'loopback')
+    }
+    const hostname = parse(baseUrl).hostname
 
+    /**
+     *  Authentication setup
+     */
     let authMiddleware = (req: any, res: any, next: any) => {
       next()
     }
     let securityScheme
     if (options.apiKey) {
+      console.log('🔐 Secured by API Key')
       passport.use(
         new Bearer.Strategy((token, done) => {
           if (!options.apiKey || options.apiKey === token) {
@@ -47,8 +65,13 @@ program
       securityScheme = 'bearer'
     }
 
+    /**
+     * Exposing agent methods
+     */
+    const exposedMethods = options.exposedMethods ? options.exposedMethods : agent.availableMethods()
     const getAgentForRequest = async (req: express.Request) => agent
 
+    console.log('🗺  API base path', baseUrl + options.apiBasePath)
     app.use(
       options.apiBasePath,
       authMiddleware,
@@ -58,9 +81,15 @@ program
       }),
     )
 
+    /**
+     * Exposing OpenAPI schema
+     */
+    console.log('🗺  OpenAPI schema', baseUrl + options.schemaPath)
+    console.log('🧩 Available methods', agent.availableMethods().length)
+    console.log('🛠  Exposed methods', exposedMethods.length)
     app.use(
       options.schemaPath,
-      AgentApiSchemaRouter({
+      ApiSchemaRouter({
         basePath: options.apiBasePath,
         getAgentForRequest,
         exposedMethods,
@@ -68,95 +97,92 @@ program
       }),
     )
 
-    app.use(AgentDidDocRouter({ getAgentForRequest }))
+    /**
+     * Handling 'did:web' requests ('/.well-known/did.json' and '/^\/(.+)\/did.json$/')
+     */
+    console.log('📋 DID Document ' + baseUrl + didDocEndpoint)
+    app.use(DidDocRouter({ getAgentForRequest }))
+
+    console.log('📖 API Documentation', baseUrl + options.apiDocsPath)
+    app.use(
+      options.apiDocsPath,
+      swaggerUi.serve,
+      swaggerUi.setup(undefined, {
+        swaggerOptions: {
+          url: baseUrl + options.schemaPath,
+        },
+      }),
+    )
+
+    /**
+     * Creating server identity and configuring messaging service endpoint
+     */
+    let serverIdentity: IIdentity
+    if (options.defaultIdentity.create) {
+      serverIdentity = await agent.identityManagerGetOrCreateIdentity({
+        provider: 'did:web',
+        alias: hostname,
+      })
+      console.log('🆔', serverIdentity.did)
+
+      const messagingServiceEndpoint = baseUrl + options.defaultIdentity.messagingServiceEndpoint
+
+      console.log('📨 Messaging endpoint', messagingServiceEndpoint)
+      await agent.identityManagerAddService({
+        did: serverIdentity.did,
+        service: {
+          id: serverIdentity.did + '#msg',
+          type: 'Messaging',
+          description: 'Handles incoming POST messages',
+          serviceEndpoint: messagingServiceEndpoint,
+        },
+      })
+
+      app.post(
+        options.defaultIdentity.messagingServiceEndpoint,
+        express.text({ type: '*/*' }),
+        async (req, res) => {
+          try {
+            const message = await agent.handleMessage({ raw: req.body, save: true })
+            console.log('Received message', message.type, message.id)
+            res.json({ id: message.id })
+          } catch (e) {
+            console.log(e)
+            res.send(e.message)
+          }
+        },
+      )
+    }
+
+    /**
+     * Serving homepage
+     */
+    app.get('/', async (req, res) => {
+      const links = [
+        { label: 'API Docs', url: options.apiDocsPath },
+        { label: 'API Schema', url: options.schemaPath },
+        { label: 'DID Document', url: didDocEndpoint },
+      ]
+
+      let verifiablePresentation
+      if (serverIdentity) {
+        const presentations = await agent.dataStoreORMGetVerifiablePresentations({
+          where: [
+            { column: 'holder', value: [serverIdentity.did] },
+            { column: 'verifier', value: [baseUrl] },
+          ],
+        })
+
+        verifiablePresentation =
+          presentations.length > 0 ? presentations[presentations.length - 1].verifiablePresentation : null
+      }
+
+      const template = options.homePageTemplate || __dirname + '/../views/home.html'
+      const rendered = await hbs.render(template, { verifiablePresentation, links })
+      res.send(rendered)
+    })
 
     app.listen(cmd.port || options.port, async () => {
-      console.log(`🚀 Agent server ready at http://localhost:${cmd.port || options.port}`)
-      console.log('🧩 Available methods', agent.availableMethods().length)
-      console.log('🛠  Exposed methods', exposedMethods.length)
-
-      let baseUrl = options.baseUrl
-
-      if (options.ngrok?.connect) {
-        baseUrl = await ngrok.connect({
-          addr: cmd.port || options.port,
-          subdomain: options.ngrok.subdomain,
-          region: options.ngrok.region,
-          authtoken: options.ngrok.authtoken,
-        })
-        app.set('trust proxy', 'loopback')
-      }
-      const hostname = parse(baseUrl).hostname
-
-      app.use(
-        options.apiDocsPath,
-        swaggerUi.serve,
-        swaggerUi.setup(undefined, {
-          swaggerOptions: {
-            url: baseUrl + options.schemaPath,
-          },
-        }),
-      )
-      console.log('📖 API Documentation', baseUrl + options.apiDocsPath)
-      console.log('🗺  OpenAPI schema', baseUrl + options.schemaPath)
-
-      if (options.defaultIdentity.create) {
-        let serverIdentity = await agent.identityManagerGetOrCreateIdentity({
-          provider: 'did:web',
-          alias: hostname,
-        })
-        console.log('🆔', serverIdentity.did)
-
-        const messagingServiceEndpoint = baseUrl + options.defaultIdentity.messagingServiceEndpoint
-
-        await agent.identityManagerAddService({
-          did: serverIdentity.did,
-          service: {
-            id: serverIdentity.did + '#msg',
-            type: 'Messaging',
-            description: 'Handles incoming POST messages',
-            serviceEndpoint: messagingServiceEndpoint,
-          },
-        })
-        console.log('📨 Messaging endpoint', messagingServiceEndpoint)
-
-        app.post(
-          options.defaultIdentity.messagingServiceEndpoint,
-          express.text({ type: '*/*' }),
-          async (req, res) => {
-            try {
-              const message = await agent.handleMessage({ raw: req.body, save: true })
-              console.log('Received message', message.type, message.id)
-              res.json({ id: message.id })
-            } catch (e) {
-              console.log(e)
-              res.send(e.message)
-            }
-          },
-        )
-
-        console.log('📋 DID Document ' + baseUrl + didDocEndpoint)
-
-        app.get('/', async (req, res) => {
-          const links = [
-            { label: 'API Docs', url: options.apiDocsPath },
-            { label: 'API Schema', url: options.schemaPath },
-            { label: 'DID Document', url: didDocEndpoint },
-          ]
-
-          const presentations = await agent.dataStoreORMGetVerifiablePresentations({
-            where: [
-              { column: 'holder', value: [serverIdentity.did] },
-              { column: 'verifier', value: [baseUrl] },
-            ],
-          })
-
-          const verifiablePresentation =
-            presentations.length > 0 ? presentations[presentations.length - 1].verifiablePresentation : null
-          const template = options.homePageTemplate || __dirname + '/../views/home.html'
-          const rendered = await hbs.render(template, { verifiablePresentation, links })
-          res.send(rendered)
-        })
-      }
+      console.log(`🚀 Cloud Agent ready at ${baseUrl}`)
     })
   })
